created reducer for search query
