--- src/components/LandingPage/LandingPageInfo.jsx
@@ -1,9 +1,0 @@
-import { Stack } from '@mui/material';
-
-function LandingPageInfo() {
-    
-    return (<>
-        <h1>THIS IS INFO</h1>
-    </>)
-}
-export default LandingPageInfo;
